Refactored code moved FeedImage mapping logic to its own helper class file
